Added gtsrb dataset support

+++ run_ssd_live_demo.py
@@ -71,7 +71,7 @@ while True:
     interval = timer.end()
     print('Time: {:.2f}s, Detect Objects: {:d}.'.format(interval, labels.size(0)))
     for i in range(boxes.size(0)):
-        box = boxes[i, :]
+        box = boxes[i, :].int()
         label = f"{class_names[labels[i]]}: {probs[i]:.2f}"
         cv2.rectangle(orig_image, (box[0], box[1]), (box[2], box[3]), (255, 255, 0), 4)
